metavision-master-main/pyproject.toml
[project]
name = "metavision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=3.0.4",
    "flask-caching>=2.3.1",
    "matplotlib>=3.10.3",
    "nibabel>=5.3.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pyinstaller>=6.13.0",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
]
